api-service/requirements.txt
aniso8601==9.0.1
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
greenlet==1.1.0
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pkg-resources==0.0.0
psycopg2==2.9.1
PyJWT==1.4.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
Werkzeug==2.0.1
xlrd==2.0.1
zipp==3.4.1
